+++ x/staking/client/cli/cli_test.go
@@ -1,4 +1,4 @@
-// build norace
+// +build norace
 
 package cli_test